chore: migrated filters.ts to packages
