+++ packages/i18n/src/locales/en/translations.json
@@ -316,5 +316,8 @@
   "change_parent_issue": "Change parent issue",
   "remove_parent_issue": "Remove parent issue",
   "add_parent": "Add parent",
-  "loading_members": "Loading members..."
+  "loading_members": "Loading members...",
+  "common": {
+    "filters": "Filters"
+  }
 }
