+++ packages/i18n/src/locales/es/translations.json
@@ -316,5 +316,8 @@
   "remove_parent_issue": "Eliminar problema padre",
   "add_parent": "Agregar padre",
   "loading_members": "Cargando miembros...",
-  "inbox": "bandeja de entrada"
+  "inbox": "bandeja de entrada",
+  "common": {
+    "filters": "Filtros"
+  }
 }
