+++ packages/i18n/src/locales/fr/translations.json
@@ -316,5 +316,8 @@
   "remove_parent_issue": "Supprimer le problème parent",
   "add_parent": "Ajouter un parent",
   "loading_members": "Chargement des membres...",
-  "inbox": "Boîte de réception"
+  "inbox": "Boîte de réception",
+  "common": {
+    "filters": "Filtres"
+  }
 }
