+++ packages/i18n/src/locales/ja/translations.json
@@ -316,5 +316,8 @@
   "remove_parent_issue": "親問題を削除",
   "add_parent": "親問題を追加",
   "loading_members": "メンバーを読み込んでいます...",
-  "inbox": "受信箱"
+  "inbox": "受信箱",
+  "common": {
+    "filters": "フィルター"
+  }
 }
