+++ packages/i18n/src/locales/zh-CN/translations.json
@@ -315,5 +315,8 @@
   "remove_parent_issue": "移除父问题",
   "add_parent": "添加父问题",
   "loading_members": "正在加载成员...",
-  "inbox": "收件箱"
+  "inbox": "收件箱",
+  "common": {
+    "filters": "过滤器"
+  }
 }
