+++ web/core/components/common/applied-filters/date.tsx
@@ -1,8 +1,8 @@
 import { observer } from "mobx-react";
 // icons
 import { X } from "lucide-react";
-// constants
-import { DATE_BEFORE_FILTER_OPTIONS } from "@/constants/filters";
+// plane constants
+import { DATE_BEFORE_FILTER_OPTIONS } from "@plane/constants";
 // helpers
 import { renderFormattedDate } from "@/helpers/date-time.helper";
 import { capitalizeFirstLetter } from "@/helpers/string.helper";
