+++ web/core/components/issues/issue-layouts/filters/applied-filters/date.tsx
@@ -2,7 +2,7 @@ import { observer } from "mobx-react";
 // icons
 import { X } from "lucide-react";
 // helpers
-import { DATE_AFTER_FILTER_OPTIONS } from "@/constants/filters";
+import { DATE_AFTER_FILTER_OPTIONS } from "@plane/constants";
 import { renderFormattedDate } from "@/helpers/date-time.helper";
 import { capitalizeFirstLetter } from "@/helpers/string.helper";
 // constants
