+++ web/core/components/issues/issue-layouts/filters/header/filters/target-date.tsx
@@ -1,11 +1,10 @@
 import React, { useState } from "react";
 import { observer } from "mobx-react";
-
+// constants
+import { DATE_AFTER_FILTER_OPTIONS } from "@plane/constants";
 // components
 import { DateFilterModal } from "@/components/core";
 import { FilterHeader, FilterOption } from "@/components/issues";
-// constants
-import { DATE_AFTER_FILTER_OPTIONS } from "@/constants/filters";
 
 type Props = {
   appliedFilters: string[] | null;
