+++ web/core/components/modules/applied-filters/date.tsx
@@ -1,8 +1,8 @@
 import { observer } from "mobx-react";
 // icons
 import { X } from "lucide-react";
+import { DATE_AFTER_FILTER_OPTIONS } from "@plane/constants";
 // helpers
-import { DATE_AFTER_FILTER_OPTIONS } from "@/constants/filters";
 import { renderFormattedDate } from "@/helpers/date-time.helper";
 import { capitalizeFirstLetter } from "@/helpers/string.helper";
 // constants
