+++ web/core/components/pages/header/root.tsx
@@ -1,6 +1,7 @@
 import { useCallback } from "react";
 import { observer } from "mobx-react";
 import { ListFilter } from "lucide-react";
+import { useTranslation } from "@plane/i18n";
 import { TPageFilterProps, TPageNavigationTabs } from "@plane/types";
 // components
 import { Header, EHeaderVariant } from "@plane/ui";
@@ -25,6 +26,7 @@ type Props = {
 
 export const PagesListHeaderRoot: React.FC<Props> = observer((props) => {
   const { pageType, projectId, workspaceSlug } = props;
+  const { t } = useTranslation();
   // store hooks
   const { filters, updateFilters, clearAllFilters } = useProjectPages();
   const {
@@ -69,7 +71,7 @@ export const PagesListHeaderRoot: React.FC<Props> = observer((props) => {
           />
           <FiltersDropdown
             icon={<ListFilter className="h-3 w-3" />}
-            title="Filters"
+            title={t("common.filters")}
             placement="bottom-end"
             isFiltersApplied={isFiltersApplied}
           >
